Check for the theme-public action in deploy-build.sh

+++ deploy-build.sh
@@ -29,15 +29,22 @@ git checkout $1
 find "/var/repos/$2" -type d -exec chmod 775 {} \;
 find "/var/repos/$2" -type f -exec chmod 664 {} \;
 
-# Remove the old theme directory if it exists.
-if [ -d "/var/repos/wsuwp-platform/build-themes/private/$2" ]; then
-  rm -fr "/var/repos/wsuwp-platform/build-themes/private/$2"
-fi
+# Our default action, a public theme being deployed by the team, requires
+# that we pull the tagged version of the theme and move it to the private
+# directory for full build and deploy. Private in this case means it is
+# not part of the collective public build of themes, but is a standalone
+# of any sort.
+if [ 'theme-public' == $4 ]; then
+  # Remove the old theme directory if it exists.
+  if [ -d "/var/repos/wsuwp-platform/build-themes/private/$2" ]; then
+    rm -fr "/var/repos/wsuwp-platform/build-themes/private/$2"
+  fi
 
-# Copy over the new theme directory and remove its .git directory. We are
-# unable to use rsync for this due to some restrictions on the server.
-cp -r "/var/repos/$2" "/var/repos/wsuwp-platform/build-themes/private/$2"
-rm -rf "/var/repos/wsuwp-platform/build-themes/private/$2/.git"
+  # Copy over the new theme directory and remove its .git directory. We are
+  # unable to use rsync for this due to some restrictions on the server.
+  cp -r "/var/repos/$2" "/var/repos/wsuwp-platform/build-themes/private/$2"
+  rm -rf "/var/repos/wsuwp-platform/build-themes/private/$2/.git"
+fi
 
 # Build the project to prep for deployment.
 cd /var/repos/wsuwp-platform/
